Atomic Design del feature home terminado
Se termino de implementar el diseño de la UI con el patrón Atomic Design, para la vista de Home(HomePage.tsx) de la app CRM
